LegacyModal is now Modal
