+++ components/src/modals/__tests__/Modal.test.tsx
@@ -3,9 +3,9 @@ import * as React from 'react'
 import '@testing-library/jest-dom/vitest'
 import { screen } from '@testing-library/react'
 import { describe, it, expect, beforeEach } from 'vitest'
-import { renderWithProviders } from '../../../testing/utils'
-import { COLORS } from '../../../helix-design-system'
-import { Modal } from '..'
+import { renderWithProviders } from '../../testing/utils'
+import { COLORS } from '../../helix-design-system'
+import { Modal } from '../index'
 
 const render = (props: React.ComponentProps<typeof Modal>) => {
   return renderWithProviders(<Modal {...props} />)
