+++ components/src/modals/__tests__/ModalHeader.test.tsx
@@ -3,11 +3,11 @@ import { screen, fireEvent } from '@testing-library/react'
 import '@testing-library/jest-dom/vitest'
 import { describe, it, expect, vi, beforeEach } from 'vitest'
 
-import { renderWithProviders } from '../../../testing/utils'
+import { renderWithProviders } from '../../testing/utils'
 import { ModalHeader } from '../ModalHeader'
-import { COLORS } from '../../../helix-design-system'
-import { SPACING } from '../../../ui-style-constants'
-import { ALIGN_CENTER, JUSTIFY_CENTER } from '../../../styles'
+import { COLORS } from '../../helix-design-system'
+import { SPACING } from '../../ui-style-constants'
+import { ALIGN_CENTER, JUSTIFY_CENTER } from '../../styles'
 
 const mockClose = vi.fn()
