--- components/src/modals/index.ts
@@ -1,10 +1,0 @@
-// modal components
-
-export * from './AlertModal'
-export * from './ContinueModal'
-export * from './Modal'
-export * from './ShameModal'
-export * from './ShameModalShell'
-export * from './Overlay'
-export * from './SpinnerModal'
-export * from './SpinnerModalPage'
